Fix bug in RGB refinement
Sorting the edges in RGB according to length only affected edge marking,
not the actual refinement step. This led to wrong refinements if the
longest edge was not the refiment edge.

Reordering edges is now outsourced to a method of mesh to do it in a
global and sustainable way. This also cleans up RGB a little.